src/app/requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.3  # PostgreSQL adapter for Python (if using PostgreSQL DB)
Werkzeug==2.2.3
